requirements.txt
Pillow==10.4.0
streamlit==1.33.0
